// src/interfaces/transitoria.interface.ts
export interface Transitoria {
    uuid?: string;
    numero_autorizacion?: string;
    vigencia?: string;
    fecha_expedicion?: string;
    // licencia_funcionamiento?: number
    nombre_comercial?: string
    giro_comercial?: string
    predio_id?: number
    nombre?: string
    apellidos?: string
    razon_social?: string
    rfc?: string
    clave_persona?: string
    direccion?: string
    codigo_postal?: string
    calle?: string
    colonia?: string
    numero_exterior?: string
    numero_interior?: string
}